+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### v0.0.10
 - Allow override of repo name in `dockci.yaml` #397
 - Messages when images will and won't be cleaned up #397
+- Login, and register with GitHub and GitLab OAuth #398
+- Configuration to turn off internal DockCI users #398
 
 ### v0.0.9
 - Exception reported to Rollbar on git push hook `ValidationError` #322